working on Dijkstra's. I think the bulk of it is done... need to follow paths and check it though

+++ gmrGraphs/Vertex.cs
@@ -15,7 +15,7 @@ namespace gmrGraphs
 
 
         public bool visited = false;
-        public float knownDistance = 0;
+        public double knownDistance = 0;
         public Vertex<T> founder;
